Add support for merge and add specs
